Now supports first validations
The program can now validate aces, twos, threes, fours, fives, sixes,
three of a kind and four of a kind. That means, it can look if you could
make a "three of a kind" with e.g. the dices [6,6,6,2,3]. Full House,
Small Straight, Long Staright, Kniffel and Chance will follow.